requirements.txt
fastapi==0.66.0
uvicorn==0.13.4
aiodns==2.0.0
aiohttp==3.7.4
photon-pump==0.9.2
python-dotenv==0.13.0